requirements.txt
torchrl
torch
tensordict
hydra-core
tensorboard
pygame
imageio
tqdm
wandb
matplotlib